test(flash_model): complete the Python suite, and correct AGENTS.md
379 pytest cases in 0.29 s. The test harness follows the DIRECTIVE STREAM
rather than assuming a command's phase structure, so lane widths and dummy
prefixes are genuinely covered instead of only the resulting bytes -- the
same reason the VHDL side checks bit order against a raw bus trace.

Adds `ignored_command_count` as the sum of the seven granular reject
counters, which tb_flash_model and doc/flash_model.md both reference.

AGENTS.md claimed the VUnit pin "carries the `--wave` flag that headless
waveform recording needs". That stopped being true when the pin moved to the
bridge fork, and a stale environment note is worse than none -- it sends the
next reader looking for a capability that is gone. Now states the trade
plainly, says reconciling the forks is the fix, and warns against silently
repinning (which would break flash_model). Also records that the bridge
compiles its native library at SIMULATION time, which is why CI needs
python3-dev and gcc.

Model behaviours that the requirements left open, now decided and documented
in-code: a trailing partial byte aborts data-in commands only and leaves WEL
set (the instruction never executed); a protection rejection clears WEL and
does not go busy; protection is any-overlap, so a write straddling a lock is
refused entirely; written_regions covers device programs and erases but not
preload; flash_reset clears volatile state and keeps both the array and the
testbench's lock map.

Verified after the change: 47 of 47 on GHDL and on NVC.

+++ modules/flash_model/python/flash_model/device.py
@@ -600,7 +600,7 @@ class FlashDevice:
             "sr1": lambda: self.status_byte(0),
             "sr2": lambda: self.status_byte(1),
             "sr3": lambda: self.status_byte(2),
-            "busy_deadline_ps": lambda: int(round(self.timing.deadline() * 1e12)),
+            "busy_deadline_ps": lambda: round(self.timing.deadline() * 1e12),
             "timing_enabled": lambda: int(self.timing.enabled),
             "materialized_pages": lambda: self.array.materialized_pages,
             "run_count": lambda: self.array.run_count,
